shifts tower metrics DAG logic and variables to separate location
